requirements.txt
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
idna==2.10
numpy==1.19.5
python-dateutil==2.8.1
python-decouple==3.4
pytz==2020.5
requests==2.25.1
six==1.15.0
soupsieve==2.1
urllib3==1.26.2
